Update paths to moved TopbarSearchForm
