+++ src/components/Topbar/Topbar.js
@@ -19,10 +19,10 @@ import {
   ModalMissingInformation,
   NamedLink,
 } from '../../components';
-import { TopbarSearchForm } from '../../forms';
 
 import MenuIcon from './MenuIcon';
 import SearchIcon from './SearchIcon';
+import TopbarSearchForm from './TopbarSearchForm/TopbarSearchForm';
 import TopbarMobileMenu from './TopbarMobileMenu/TopbarMobileMenu';
 import TopbarDesktop from './TopbarDesktop/TopbarDesktop';
