+++ src/components/Topbar/TopbarSearchForm/TopbarSearchForm.module.css
@@ -1,4 +1,4 @@
-@import '../../styles/propertySets.css';
+@import '../../../styles/propertySets.css';
 
 .mobileInputRoot {
   width: 100%;
