+++ src/forms/index.js
@@ -1,4 +1,3 @@
 export { default as BookingDatesForm } from './BookingDatesForm/BookingDatesForm';
 export { default as StripePaymentForm } from './StripePaymentForm/StripePaymentForm';
 export { default as StripeConnectAccountForm } from './StripeConnectAccountForm/StripeConnectAccountForm';
-export { default as TopbarSearchForm } from './TopbarSearchForm/TopbarSearchForm';
